+++ CHANGELOG.rst
@@ -5,6 +5,14 @@ All notable changes to this project will be documented in this file.
 The format is based on `Keep a Changelog <https://keepachangelog.com/en/1.1.0/>`__,
 and this project adheres to `Semantic Versioning <(https://semver.org/spec/v2.0.0.html>`__.
 
+Unreleased
+----------
+
+Added
+^^^^^
+
+- ``read_file_formats.file_formats`` now accepts ``data`` parameter to parse user-provided list instead of the pre-packaged data
+
 1.1.1 - 2025-10-23
 ------------------